chore(refactor): extract /api/auth/* into routes/auth.ts
Fifth bucket of the index.ts route extraction. Three endpoints:
  GET  /api/auth/whoami
  POST /api/auth/login   (starts the device-flow sign-in)
  POST /api/auth/logout

The local routes are just glue around AuthService — the actual auth
provider (magic-link, OAuth) lives in infra/auth-worker/.

Drive-by cleanup: the /api/ui/events comment block was sitting one block
above its handler in index.ts (separated by the auth blocks during some
earlier reorder). Moved back to where it documents what it documents.

No behavioural change.
